cpu: Fix SwSpan uint16_t overflow for large canvases
SwSpan used uint16_t for x, y, and len fields, which overflows at
65535 pixels. At 2x scale on large documents (e.g. sheet music scores),
pixel dimensions easily exceed this limit, causing rendering corruption.

Promote SwSpan fields to uint32_t and update all cast sites in
tvgSwRle.cpp accordingly.

+++ src/renderer/cpu_engine/tvgSwCommon.h
@@ -114,8 +114,8 @@ struct SwOutline
 
 struct SwSpan
 {
-    uint16_t x, y;
-    uint16_t len;
+    int32_t x, y;
+    int32_t len;
     uint8_t coverage;
 
     bool fetch(const RenderRegion& bbox, int32_t& x, int32_t& len) const
@@ -135,7 +135,7 @@ struct SwRle
         return fetch(bbox.min.y, bbox.max.y - 1, end);
     }
 
-    const SwSpan* fetch(int32_t min, uint32_t max, const SwSpan** end) const
+    const SwSpan* fetch(int32_t min, int32_t max, const SwSpan** end) const
     {
         const SwSpan* begin;
